aligh with release/12.2.0

+++ sys/ofed/include/rdma/ib_verbs.h
@@ -37,7 +37,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  *
- * $FreeBSD: releng/12.1/sys/ofed/include/rdma/ib_verbs.h 347766 2019-05-16 16:27:21Z hselasky $
+ * $FreeBSD$
  */
 
 #if !defined(IB_VERBS_H)
@@ -285,7 +285,7 @@ enum ib_cq_creation_flags {
 
 struct ib_cq_init_attr {
 	unsigned int	cqe;
-	int		comp_vector;
+	u32		comp_vector;
 	u32		flags;
 };
 
@@ -1559,7 +1559,7 @@ struct ib_mr {
 	u32		   lkey;
 	u32		   rkey;
 	u64		   iova;
-	u32		   length;
+	u64		   length;
 	unsigned int	   page_size;
 	bool		   need_inval;
 	union {
